Check for .cid exists before trying to rm it

+++ build.sh
@@ -4,7 +4,9 @@ set -e
 from=nowk/alpine-nodejs:0.12.2
 
 # remove .cid, incase there is a lingering one
-rm .cid
+if [ -f .cid ] ; then
+	rm .cid
+fi
 
 docker build --rm -t $from .
 docker run --privileged=true --cidfile .cid $from /max_user_watches.sh
